Alligator captured in Brooklyn's Prospect Park Lake A 4-foot-long alligator was pulled out of Prospect Park Lake on Sunday after being spotted earlier in the day by park workers, according to media reports.
Nobody was injured by the animal, which lethargic and in poor condition. It was taken to the Bronx Zoo, the reports said.

The northernmost part of the American alligator’s natural habitat is North Carolina, and they can grow up to 12 feet in length, according to The National Wildlife Federation.

Parks officials told NY1 that they had named the alligator “Godzilla.”

It is not the first nonnative reptile found in city parks.

In June 2001, an off duty police officer and his two children spotted what they thought was an alligator in the Harlem Meer. That sighting started a citywide hunt for the animal. Five days after the first sighting, a 2-foot spectacled caiman was captured. It became known as “Daimon the Caiman.”

A few years later, in 2015, an alligator nicknamed “Jaws” was hunted for weeks before it was captured at Flushing Meadow Corona Park.

It seems alligator stories have been with us for a while, though. It was 88 years ago this month when, the story goes, a Harlem teenager spotted an alligator in a sewer, giving birth to one of the most enduring urban legends of all time.